+++ src/index.js
@@ -119,6 +119,9 @@ import {
   CATEGORICAL,
   VALUE_ZW_DATA_TYPES,
   DEFAULT_SPATIAL_INDEX_USE_WORKER,
+  DEFAULT_ANNOTATION_LINE_COLOR,
+  DEFAULT_ANNOTATION_LINE_WIDTH,
+  DEFAULT_ANNOTATION_HVLINE_LIMIT,
 } from './constants';
 
 
@@ -143,6 +146,11 @@ import {
   rgbBrightness,
   clip,
   toArrayOrientedPoints,
+  isHorizontalLine,
+  isVerticalLine,
+  isDomRect,
+  isRect,
+  isPolygon,
 } from './utils';
 
 import { version } from '../package.json';
@@ -263,6 +271,9 @@ const createScatterplot = (
     sizeBy = DEFAULT_SIZE_BY,
     height = DEFAULT_HEIGHT,
     width = DEFAULT_WIDTH,
+    annotationLineColor = DEFAULT_ANNOTATION_LINE_COLOR,
+    annotationLineWidth = DEFAULT_ANNOTATION_LINE_WIDTH,
+    annotationHVLineLimit = DEFAULT_ANNOTATION_HVLINE_LIMIT,
   } = initialProperties;
 
   let currentWidth = width === AUTO ? 1 : width;
@@ -291,7 +302,10 @@ const createScatterplot = (
   let isDestroyed = false;
   let backgroundColorBrightness = rgbBrightness(backgroundColor);
   let camera;
+  /** @type {ReturnType<createLine>} */
   let lasso;
+  /** @type {ReturnType<createLine>} */
+  let annotations;
   let mouseDown = false;
   let mouseDownTime = null;
   let mouseDownPosition = [0, 0];
@@ -473,6 +487,7 @@ const createScatterplot = (
 
   let isViewChanged = false;
   let isPointsDrawn = false;
+  let isAnnotationsDrawn = false;
   let isMouseOverCanvasChecked = false;
 
   let valueZDataType = CATEGORICAL;
@@ -2412,13 +2427,141 @@ const createScatterplot = (
     );
   };
 
-  /** @type {<F extends Function>(f: F) => (...args: Parameters<F>) => ReturnType<F>} */
+  /**
+   * Draw line-based annotations.
+   * @param {import('./types').Annotation[]} newAnnotations
+   * @returns {Promise<void>}
+   */
+  const drawAnnotations = (newAnnotations) => {
+    if (isDestroyed) {
+      return Promise.reject(new Error('The instance was already destroyed'));
+    }
+
+    isAnnotationsDrawn = false;
+
+    if (newAnnotations.length === 0) {
+      return new Promise((resolve) => {
+        annotations.clear();
+        pubSub.subscribe('draw', resolve, 1);
+        isAnnotationsDrawn = true;
+        draw = true;
+      });
+    }
+
+    return new Promise((resolve) => {
+      const newPoints = [];
+      const newColors = new Map();
+      const newColorIndices = [];
+      const newWidths = [];
+
+      let maxNewColorIdx = -1;
+
+      const addColorAndWidth = (annotation) => {
+        newWidths.push(annotation.lineWidth || annotationLineWidth);
+
+        const color = toRgba(annotation.lineColor || annotationLineColor, true);
+        const colorId = `[${color.join(',')}]`;
+        if (newColors.has(colorId)) {
+          const { idx } = newColors.get(colorId);
+          newColorIndices.push(idx);
+        } else {
+          const idx = ++maxNewColorIdx;
+          newColors.set(colorId, { idx, color });
+          newColorIndices.push(idx);
+        }
+      };
+
+      for (const annotation of newAnnotations) {
+        if (isHorizontalLine(annotation)) {
+          newPoints.push([
+            annotation.x1 ?? -annotationHVLineLimit,
+            annotation.y,
+            annotation.x2 ?? annotationHVLineLimit,
+            annotation.y,
+          ]);
+          addColorAndWidth(annotation);
+          continue;
+        }
+
+        if (isVerticalLine(annotation)) {
+          newPoints.push([
+            annotation.x,
+            annotation.y1 ?? -annotationHVLineLimit,
+            annotation.x,
+            annotation.y2 ?? annotationHVLineLimit,
+          ]);
+          addColorAndWidth(annotation);
+          continue;
+        }
+
+        if (isRect(annotation)) {
+          newPoints.push([
+            annotation.x1,
+            annotation.y1,
+            annotation.x2,
+            annotation.y1,
+            annotation.x2,
+            annotation.y2,
+            annotation.x1,
+            annotation.y2,
+            annotation.x1,
+            annotation.y1,
+          ]);
+          addColorAndWidth(annotation);
+          continue;
+        }
+
+        if (isDomRect(annotation)) {
+          newPoints.push([
+            annotation.x,
+            annotation.y,
+            annotation.x + annotation.width,
+            annotation.y,
+            annotation.x + annotation.width,
+            annotation.y + annotation.height,
+            annotation.x,
+            annotation.y + annotation.height,
+            annotation.x,
+            annotation.y,
+          ]);
+          addColorAndWidth(annotation);
+          continue;
+        }
+
+        if (isPolygon(annotation)) {
+          newPoints.push(annotation.vertices.flatMap(identity));
+          addColorAndWidth(annotation);
+        }
+      }
+
+      annotations.setStyle({
+        color: Array.from(newColors.values())
+          .sort((a, b) => (a.idx > b.idx ? 1 : -1))
+          .map(({ color }) => color),
+      });
+      annotations.setPoints(
+        newPoints.length === 1 ? newPoints.flat() : newPoints,
+        {
+          colorIndices: newColorIndices,
+          widths: newWidths,
+        }
+      );
+
+      pubSub.subscribe('draw', resolve, 1);
+      isAnnotationsDrawn = true;
+      draw = true;
+    });
+  };
+
+  /** @type {<F extends Function>(f: F) => (...args: Parameters<F>) => Promise<ReturnType<F>>} */
   const withDraw =
     (f) =>
     (...args) => {
       const out = f(...args);
       draw = true;
-      return out;
+      return new Promise((resolve) => {
+        pubSub.subscribe('draw', () => resolve(out), 1);
+      });
     };
 
   /**
@@ -2445,25 +2588,33 @@ const createScatterplot = (
 
   /**
    * Zoom to an area specified as a rectangle
-   * @param {import('./types').Rect} rect - The rectangle to zoom to
-   * @param {import('./types').ScatterplotMethodOptions['draw']} options
+   * @param {import('./types').Rect} rect - The rectangle to zoom to in normalized device coordinates
+   * @param {import('./types').ScatterplotMethodOptions['zoomToArea']} options
    * @returns {Promise<void>}
    */
   const zoomToArea = (rect, options = {}) =>
     new Promise((resolve) => {
-      const target = [rect.x + rect.width / 2, rect.y + rect.height / 2];
+      const target = vec4
+        .transformMat4(
+          [],
+          [rect.x + rect.width / 2, rect.y + rect.height / 2, 0, 0],
+          model
+        )
+        .slice(0, 2);
 
       // Vertical field of view
       // The Arc Tangent is based on the original camera position. Otherwise
       // we would have to do `Math.atan(1 / camera.view[5])`
       const vFOV = 2 * Math.atan(1);
 
+      const aspectRatio = viewAspectRatio / dataAspectRatio;
+
       const distance =
-        rect.height * viewAspectRatio > rect.width
+        rect.height * aspectRatio >= rect.width
           ? // Distance is based on the height of the bounding box
             rect.height / 2 / Math.tan(vFOV / 2)
           : // Distance is based on the width of the bounding box
-            rect.width / 2 / Math.tan((vFOV * viewAspectRatio) / 2);
+            rect.width / 2 / Math.tan(vFOV / 2) / aspectRatio;
 
       if (options.transition) {
         camera.config({ isFixed: true });
@@ -2513,9 +2664,9 @@ const createScatterplot = (
 
   /**
    * Zoom to a location specified in normalized devide coordinates.
-   * @param {number[]} target - The camera target
+   * @param {number[]} target - The camera target given in normalized device coordinates
    * @param {number} distance - The camera distance
-   * @param {import('./types').ScatterplotMethodOptions['draw']} options
+   * @param {import('./types').ScatterplotMethodOptions['zoomToLocation']} options
    * @returns {Promise<void>}
    */
   const zoomToLocation = (target, distance, options = {}) =>
@@ -2544,7 +2695,7 @@ const createScatterplot = (
 
   /**
    * Zoom to the origin
-   * @param {import('./types').ScatterplotMethodOptions['draw']} options
+   * @param {import('./types').ScatterplotMethodOptions['zoomToLocation']} options
    * @returns {Promise<void>}
    */
   const zoomToOrigin = (options = {}) => zoomToLocation([0, 0], 1, options);
@@ -2931,6 +3082,18 @@ const createScatterplot = (
     opacityInactiveScale = +newOpacityInactiveScale;
   };
 
+  const setAnnotationLineColor = (newAnnotationLineColor) => {
+    annotationLineColor = toRgba(newAnnotationLineColor);
+  };
+
+  const setAnnotationLineWidth = (newAnnotationLineWidth) => {
+    annotationLineWidth = +newAnnotationLineWidth;
+  };
+
+  const setAnnotationHVLineLimit = (newAnnotationHVLineLimit) => {
+    annotationHVLineLimit = +newAnnotationHVLineLimit;
+  };
+
   const setGamma = (newGamma) => {
     renderer.gamma = newGamma;
   };
@@ -3042,9 +3205,13 @@ const createScatterplot = (
     if (property === 'isDestroyed') return isDestroyed;
     if (property === 'isPointsDrawn') return isPointsDrawn;
     if (property === 'isPointsFiltered') return isPointsFiltered;
+    if (property === 'isAnnotationsDrawn') return isAnnotationsDrawn;
     if (property === 'zDataType') return valueZDataType;
     if (property === 'wDataType') return valueWDataType;
     if (property === 'spatialIndex') return spatialIndex?.data;
+    if (property === 'annotationLineColor') return annotationLineColor;
+    if (property === 'annotationLineWidth') return annotationLineWidth;
+    if (property === 'annotationHVLineLimit') return annotationHVLineLimit;
 
     return undefined;
   };
@@ -3290,6 +3457,18 @@ const createScatterplot = (
       setGamma(properties.gamma);
     }
 
+    if (properties.annotationLineColor !== undefined) {
+      setAnnotationLineColor(properties.annotationLineColor);
+    }
+
+    if (properties.annotationLineWidth !== undefined) {
+      setAnnotationLineWidth(properties.annotationLineWidth);
+    }
+
+    if (properties.annotationHVLineLimit !== undefined) {
+      setAnnotationHVLineLimit(properties.annotationHVLineLimit);
+    }
+
     // setWidth and setHeight can be async when width or height are set to
     // 'auto'. And since draw() would have anyway been async we can just make
     // all calls async.
@@ -3383,9 +3562,25 @@ const createScatterplot = (
   };
 
   /** @type {() => void} */
-  const clear = () => {
+  const clearPoints = () => {
     setPoints([]);
     pointConnections.clear();
+  };
+
+  /** @type {() => void} */
+  const clearPointConnections = () => {
+    pointConnections.clear();
+  };
+
+  /** @type {() => void} */
+  const clearAnnotations = () => {
+    drawAnnotations([]);
+  };
+
+  /** @type {() => void} */
+  const clear = () => {
+    clearPoints();
+    clearAnnotations();
   };
 
   const resizeHandler = () => {
@@ -3438,6 +3633,11 @@ const createScatterplot = (
     reticleVLine = createLine(renderer.regl, {
       color: reticleColor,
       width: 1,
+      is2d: true,
+    });
+    annotations = createLine(renderer.regl, {
+      color: annotationLineColor,
+      width: annotationLineWidth,
       is2d: true,
     });
     computePointSizeMouseDetection();
@@ -3495,7 +3695,8 @@ const createScatterplot = (
     // Update camera: this needs to happen on every
     isViewChanged = camera.tick();
 
-    if (!isPointsDrawn || !(draw || isTransitioning)) return;
+    if (!(isPointsDrawn || isAnnotationsDrawn) || !(draw || isTransitioning))
+      return;
 
     if (isTransitioning && !tween(transitionDuration, transitionEasing))
       endTransition();
@@ -3529,10 +3730,16 @@ const createScatterplot = (
         });
       }
 
-      drawPointBodies();
+      if (isPointsDrawn) drawPointBodies();
       if (!mouseDown && (showReticle || drawReticleOnce)) drawReticle();
       if (hoveredPoint >= 0) drawHoveredPoint();
       if (selectedPoints.length) drawSelectedPoints();
+
+      annotations.draw({
+        projection: getProjection(),
+        model: getModel(),
+        view: getView(),
+      });
 
       lasso.draw({
         projection: getProjection(),
@@ -3572,6 +3779,7 @@ const createScatterplot = (
 
   const destroy = () => {
     isPointsDrawn = false;
+    isAnnotationsDrawn = false;
     isDestroyed = true;
     cancelFrameListener();
     window.removeEventListener('keyup', keyUpHandler, false);
@@ -3620,6 +3828,9 @@ const createScatterplot = (
       return renderer.isSupported;
     },
     clear: withDraw(clear),
+    clearPoints: withDraw(clearPoints),
+    clearPointConnections: withDraw(clearPointConnections),
+    clearAnnotations: withDraw(clearAnnotations),
     createTextureFromUrl: (
       /** @type {string} */ url,
       /** @type {number} */ timeout = DEFAULT_IMAGE_LOAD_TIMEOUT
@@ -3627,6 +3838,7 @@ const createScatterplot = (
     deselect,
     destroy,
     draw: publicDraw,
+    drawAnnotations,
     filter,
     get,
     getScreenPosition,