+++ src/types.d.ts
@@ -85,6 +85,48 @@ type Rect = {
   height: number;
 };
 
+interface BaseAnnotation {
+  lineColor?: Color;
+  lineWidth?: number;
+}
+
+interface AnnotationHLine extends BaseAnnotation {
+  y: number;
+  x1?: number;
+  x2?: number;
+}
+
+interface AnnotationVLine extends BaseAnnotation {
+  x: number;
+  y1?: number;
+  y2?: number;
+}
+
+interface AnnotationDomRect extends BaseAnnotation {
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+}
+
+interface AnnotationRect extends BaseAnnotation {
+  x1: number;
+  y1: number;
+  x2: number;
+  y2: number;
+}
+
+interface AnnotationPolygon extends BaseAnnotation {
+  vertices: [number, number][];
+}
+
+type Annotation =
+  | AnnotationHLine
+  | AnnotationVLine
+  | AnnotationDomRect
+  | AnnotationRect
+  | AnnotationPolygon;
+
 interface BaseOptions {
   backgroundColor: Color;
   deselectOnDblClick: boolean;
@@ -102,6 +144,9 @@ interface BaseOptions {
   width: 'auto' | number;
   gamma: number;
   aspectRatio: number;
+  annotationLineColor: Color;
+  annotationLineWidth: number;
+  annotationHVLineLimit: number;
   // Nullifiable
   backgroundImage: null | import('regl').Texture2D | string;
   colorBy: null | DataEncoding;
@@ -193,6 +238,16 @@ export interface ScatterplotMethodOptions {
   }>;
   zoomToPoints: Partial<{
     padding: number;
+    transition: boolean;
+    transitionDuration: number;
+    transitionEasing: (t: number) => number;
+  }>;
+  zoomToArea: Partial<{
+    transition: boolean;
+    transitionDuration: number;
+    transitionEasing: (t: number) => number;
+  }>;
+  zoomToLocation: Partial<{
     transition: boolean;
     transitionDuration: number;
     transitionEasing: (t: number) => number;